feat: separate use-webpack-alias from web
