+++ package.json
@@ -151,7 +151,7 @@
     "react-test-renderer": "^16.13.1",
     "resolve": "^1.15.1",
     "rimraf": "^3.0.2",
-    "rocket-punch": "^1.3.1",
+    "rocket-punch": "^1.3.3",
     "rxjs": "^6.6.0",
     "sass-loader": "^9.0.1",
     "semver": "^7.1.3",
