+++ src/@rocket-scripts/use-webpack-alias/index.ts
@@ -1,0 +1,2 @@
+export * from './getWebpackAlias';
+export * from './useWebpackAlias';
